requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.2.1
boto3==1.40.55
botocore==1.40.55
click==8.3.0
email-validator==2.2.0
fastapi==0.119.1
h11==0.16.0
idna==3.11
jmespath==1.0.1
pydantic==2.12.3
pydantic_core==2.41.4
pydantic-settings==2.5.2
pyjwt==2.10.1
pymysql==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
